requirements.txt
streamlit==1.38.0
langchain-groq==0.2.0
python-dotenv==1.0.1
pandas==2.2.3
openpyxl==3.1.5
sqlalchemy==2.0.35